Slight refactor of hud elements (this system might completely get removed...)

+++ src/main/kotlin/me/nobaboy/nobaaddons/config/ui/elements/Element.kt
@@ -4,6 +4,6 @@ data class Element(
 	val identifier: String,
 	var x: Int,
 	var y: Int,
-	var scale: Double = 1.0,
+	var scale: Float = 1.0f,
 	var color: Int = 0xFFFFFF
 )
